requirements.txt
anytree==2.12.1
bertopic==0.16.0
contextualized_topic_models==2.5.0
contractions==0.1.73
dask==2023.5.0
dspy_ai==2.4.5
gensim==4.2.0
hdbscan==0.8.33
langdetect==1.0.9
lingua==4.15.0
matplotlib==3.7.1
networkx==3.1
nltk==3.9.1
numpy==1.24.3
openai==1.44.0
pandas==2.0.3
pyarrow==16.1.0
pyfiglet==1.0.2
python-dotenv==1.0.1
pytz==2022.7
PyYAML==6.0.2
regex==2023.10.3
Requests==2.32.3
scikit_learn==1.3.2
scipy==1.10.1
sentence_transformers==3.0.1
spacy==3.7.4
spacy_download==1.1.0
tenacity==8.2.3
termcolor==2.4.0
tiktoken==0.7.0
torch==2.0.1
tqdm==4.66.4
ujson==5.8.0
umap_learn==0.5.4
Unidecode==1.3.8
lingua-language-detector==2.0.2